refactor: move initContainer
