+++ packages/client/test/test-utils/fake/fakeEnvironment.ts
@@ -5,7 +5,8 @@ import { ConfigInjectionToken, createStrictConfig, StreamrClientConfig, StrictSt
 import { DestroySignal } from '../../../src/DestroySignal'
 import { AuthConfig } from '../../../src/Ethereum'
 import { StorageNodeRegistry } from '../../../src/registry/StorageNodeRegistry'
-import { initContainer, StreamrClient } from '../../../src/StreamrClient'
+import { StreamrClient } from '../../../src/StreamrClient'
+import { initContainer } from '../../../src/Container'
 import { StreamRegistry } from '../../../src/registry/StreamRegistry'
 import { FakeBrubeckNode } from './FakeBrubeckNode'
 import { ActiveNodes } from './ActiveNodes'
