+++ packages/client/test/unit/Publisher.test.ts
@@ -3,7 +3,7 @@ import { container as rootContainer } from 'tsyringe'
 import { StreamMessage, toStreamID } from 'streamr-client-protocol'
 import { BrubeckNode, NetworkNodeStub } from '../../src/BrubeckNode'
 import { Publisher } from '../../src/publish/Publisher'
-import { initContainer } from '../../src'
+import { initContainer } from '../../src/Container'
 import { StreamRegistry } from '../../src/registry/StreamRegistry'
 import { DEFAULT_PARTITION } from '../../src/StreamIDBuilder'
 import { FakeStreamRegistry } from '../test-utils/fake/FakeStreamRegistry'
