+++ packages/client/test/unit/Stream.test.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata'
 import { container as rootContainer } from 'tsyringe'
 import { toStreamID } from 'streamr-client-protocol'
-import { initContainer } from '../../src/StreamrClient'
+import { initContainer } from '../../src/Container'
 import { Stream } from '../../src/Stream'
 import { StreamRegistry } from '../../src/registry/StreamRegistry'
 import { createStrictConfig } from '../../src/Config'
